+++ packages/unlimited/lib/src/state.dart
@@ -1,0 +1,7 @@
+import 'package:meta/meta.dart';
+import 'package:unlimited/cards.dart';
+import 'package:unlimited/schema.dart';
+import 'package:unlimited/src/internal.dart';
+
+part 'state/inventory.dart';
+part 'state/deck.dart';